% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloIMR_function.R
\name{helloIMR_function}
\alias{helloIMR_function}
\title{A helloIMR Function}
\usage{
helloIMR_function(place = "BERGEN")
}
\arguments{
\item{place}{Where are you located? Defaults to BERGEN.}
}
\description{
This function allows you to say "Hello from IMR ..."
}
\examples{
helloIMR_function()
}
\keyword{IMR}
